requirements.txt
anyio==4.9.0
asgiref==3.8.1
brotli==1.2.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
cryptography==42.0.8
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.1.5
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-model-utils==5.0.0
django-taggit==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.61.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Markdown==3.6
MarkupSafe==3.0.2
mysqlclient==2.2.6
oauthlib==3.3.1
packaging==24.2
pillow==10.3.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.12.1
PyJWT==2.10.1
pyOpenSSL==24.1.0
pyphen==0.17.2
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.3.0
requests==2.31.0
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.7.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
Werkzeug==3.0.2
wheel==0.38.1
whitenoise==6.6.0
zopfli==0.4.0
